refactoring reorganized router paths and pages directory
