+++ frontend/src/App.tsx
@@ -2,12 +2,12 @@ import React, {useState} from 'react';
 import {BrowserRouter, Route, Routes} from "react-router-dom";
 import {Layout} from "./pages/Layout";
 import {Home} from "./pages/Home";
-import {Courses} from "./pages/Courses";
-import {Teachers} from "./pages/Teachers";
+import {Courses} from "./pages/courses/Courses";
+import {Teachers} from "./pages/teachers/Teachers";
 import {NoPage} from "./pages/NoPage";
 import Login from "./pages/user/Login";
-import SingleCourse from "./pages/SingleCourse";
-import SingleTeacher from "./pages/SingleTeacher";
+import SingleCourse from "./pages/courses/SingleCourse";
+import SingleTeacher from "./pages/teachers/SingleTeacher";
 import {User} from "./lib/User";
 import {CurrentUserContext} from "./hooks/useUser";
 import Register from "./pages/user/Register";
