+++ frontend/src/pages/courses/Courses.tsx
@@ -1,20 +1,21 @@
-import CourseList from "../components/CourseList";
+import CourseList from "../../components/CourseList";
 import {useEffect, useState} from "react";
-import {Course} from "../lib/Course";
+import {Course} from "../../lib/Course";
 
 export function Courses() {
     const [courses, setCourses] = useState<Course[]>([]);
     const [isLoaded, setIsLoaded] = useState(false);
+    const [query, setQuery] = useState("");
 
     useEffect(() => {
-        fetch(`${process.env.REACT_APP_API_ROOT}courses`)
+        fetch(`${process.env.REACT_APP_API_ROOT}courses?name=${query}`) // TODO implement in backend
             .then(response => response.json())
             .then(json => {
                 setCourses(json._embedded.courses);
                 setIsLoaded(true);
             })
             .catch(e => console.error(e));
-    }, []);
+    }, [query]);
 
     const content = isLoaded
         ? <CourseList courses={courses}/>
@@ -22,6 +23,7 @@ export function Courses() {
 
     return <>
         <h1>Courses</h1>
+        <input type="text" onChange={e => setQuery(e.target.value)}/>
         {content}
     </>
 }
